Extract handler functionality into separate file
